Migrate from allprojects/subprojects to convention plugins
Replace allprojects{}, subprojects{}, and configure(filter) blocks in
the root build.gradle.kts with precompiled script plugins in buildSrc:

- webauthn4j.java-conventions: shared Java settings (compiler flags,
  test config, jacoco, reproducible archives) applied by all subprojects
- webauthn4j.java-library-conventions: extends java-conventions with
  signing and Maven publishing, applied by webauthn4j-* library modules

Each subproject now explicitly declares which convention plugin it uses
in its own plugins{} block.
