+++ webauthn4j-metadata-async/build.gradle.kts
@@ -14,6 +14,10 @@
  * limitations under the License.
  */
 
+plugins {
+    id("webauthn4j.java-library-conventions")
+}
+
 description = "WebAuthn4J Metadata Async library"
 
 dependencies {
